requirements.txt
anthropic
beautifulsoup4
pyyaml
requests
duckdb
pytest
